Add reload/swipe/timer features and eliminate all warnings
New features:
- SIGHUP handler triggers config reload (same flag as UDP __reload)
- Config file mtime watcher: checks every 2s, auto-reloads on change
- LCD swipe gestures: left = next page, right = previous (cycles main/tools/ship)
- LCD double-tap: segment 0 toggles mic mute, segment 1 cycles output device,
  segments 2-3 open Spotify when music is playing or GitHub PRs otherwise
- Timer expiry sound: plays "timer_done" via soundboard when timer hits 00:00
- Startup sound: plays "success" after boot and initial button render
- Graceful icon loading in render_button: warns instead of silently skipping

Warning/clippy cleanup:
- Fix all 24 compiler warnings (unused imports, variables, dead code, parens)
- Fix all 18 clippy lints (collapsible ifs, needless returns/borrows, single_match,
  unnecessary casts, manual_contains)
- Add unsafe block inside audio_property_changed for Rust 2024 edition compliance
- Fix min/max constant result bug in lights.rs adjust_temp
- Remove dead code: old name-based audio device switching functions
- Allow dead_code on protocol reference functions and future-use API surfaces

+++ src/boot.rs
@@ -49,7 +49,7 @@ pub fn render_boot_frame(deck: &mut StreamDeck, progress: f32, status: &str) {
     // Progress bar background
     let bar_x = 20_i32;
     let bar_y = 55_i32;
-    let bar_w = (STRIP_W - 40) as u32;
+    let bar_w = STRIP_W - 40;
     let bar_h = 20_u32;
     draw_filled_rect_mut(&mut img, Rect::at(bar_x, bar_y).of_size(bar_w, bar_h), BAR_BG);
